requirements.txt
diffusers
torch
transformers
safetensors
Pillow
streamlit
accelerate
scipy
